pyproject.toml
[tool.poetry]
name = "smarthome-mock-ai"
version = "0.1.0"
description = "一个基于 LLM 的智能家居自动化模拟系统"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "smarthome_mock_ai", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.28.1"
python-dotenv = "^1.2.1"
SpeechRecognition = "^3.10.0"
pyaudio = {version = "^0.2.13", markers = "sys_platform != 'darwin' or platform_machine != 'arm64'"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # 允许测试中使用 assert

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/smarthome_mock_ai --cov-report=term-missing"
